IG: Fix lost selector recalcs which caused stalled anchors
When a child recalcs, does not resolve, and adds a parent which was
recalculated in the same round of stabilization, the child will not get
a recalc scheduled.

The parent add always runs immediately in the next round, so we fix
this by seeing if the parent is resolved but any new children are
unresolved, and if so we schedule the children's recalculation.