Immutable graph: Effects run even when selector isn't in graph
Formerly, as an optimization, if no one depended on a selector before or
after a transaction, all commands for that selector were dropped. This
means that no effects would be generated or run.

This commit removes this optimization. Now there is a danger of
selectors flapping in an out of existence if their effects issue
commands on themselves repeatedly. On the other hand, there are no more
surprises about side effects not being run.