Drop hitch.graph/hook-next(-sel)
hook-next is seldom used (only in tests) and difficult to make behave
the same among all the kinds of graphs we have.

Eager graphs (IEagerSelectorResolve, i.e. the mutable graph) and
non-eager graphs differ about when it's best to call change-notify.
Eager graphs always expect attempt-eager-selector-resolution! to be
called immediately before an update-parents call on the same selector,
and will never call -change-notify when an external-child is added, so
hook-next would need to call its callback if its
attempt-eager-selector-resolution! call caused the selector to exist.

Immutable graphs will always call change-notify when an external child
is first added to inform it of the initial value, so hook-next would
need to detect this and drop the first notification.