Agent graph manager: always run effects, even after graph stopped
Destroy effects may run after the graph stops, so we can't check if the
graph is running. Effect fns must be able to handle a dead graph.

To ensure destroy effects always run last relative to all other effects,
we run all effects generated by a graph in a single transaction to
completion before running the next transaction's effects. This may
reduce parallelism but is safer. Otherwise we would need to track all
in-flight effects and run the graph-wide destroy effects only after we
knew the others were retired.