Fix recalc logic so child-add inform works on partially-collected node
On an immutable graph, when a node is disconnected from the graph
(orphaned, parents and children are empty) by garbage collection but the
node itself is not yet deleted (desirable for hot-caching), if that node
later regains children (becoming "alive" again) its Informed parents
or Machines may not be told about the child-addition and the selector
itself may never be recalculated.

This is an especially big problem for Machines which reset vars
based on some stateful graph-external service (e.g. graph forwarding,
rule engine) that has a sub/unsub style api. In this case the machine
won't know about its new child var, won't transmit the subscribe upward,
and won't get pertinent updates from the upstream system.